+++ src/libft/ft_putnbr_fd.c
@@ -6,7 +6,7 @@
 /*   By: shinfray <shinfray@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 21:26:06 by shinfray          #+#    #+#             */
-/*   Updated: 2023/07/31 11:48:54 by shinfray         ###   ########.fr       */
+/*   Updated: 2023/07/31 13:11:38 by shinfray         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	ft_putnbr_fd(int n, int fd)
 {
 	unsigned int	uns_n;
 
-	uns_n = n;
+	uns_n = (unsigned int)n;
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
