+++ src/libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: shinfray <shinfray@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 22:45:03 by shinfray          #+#    #+#             */
-/*   Updated: 2023/07/31 12:00:02 by shinfray         ###   ########.fr       */
+/*   Updated: 2023/07/31 13:11:19 by shinfray         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 		return ;
 	while (s[i] != '\0')
 	{
-		(*f)(i, &s[i]);
+		(*f)((unsigned int)i, &s[i]);
 		++i;
 	}
 }
